Add basic work system (#3)
## The Work System

This is a simple system that allows users to earn currency by working.
It's designed as a way to experiment with stateful applications in
bytebot using only the core features of the framework.

### State management
Normally, sane people here would use a real database to store state. But
we're not sane people.

Flow:
- user calls `!work` command
- bot handles the command and eventually routes it to `Profile.work()`
- bot calls `getProfile()` and initializes a new empty profile for the
user if one is not found
- bot passes the profile to `work()`
- bot generates a random amount of money, updates the balance, and
writes the profile back to redis
- bot returns a reply to chat indicating how much was earned and how
much the user has

At this point in time, the work system is pretty simple. Punch the
clock, get paid. No cooldown. It's a start.


![Screenshot from 2022-11-21
09-15-44](https://user-images.githubusercontent.com/2350379/203077056-4e53cb11-c5c2-4bbe-8b59-01c2dc321aba.png)